refactor(cli): split health_cmd into a command package
The health command had grown into a single ~1160-line module mixing the
Click entrypoint, table/summary rendering, refactoring-target ranking and
plan renderers, the generate-code path, trend history, and DB persistence.

Split it into a health_cmd/ package mirroring init_cmd/ and update_cmd/:

- command.py            the Click entrypoint and orchestration
- summary.py            table-mode headline sections
- refactoring_targets.py  target ranking + the twelve plan renderers
- codegen.py            the generate-code LLM path
- trends.py             trend history rendering
- persist.py            health snapshot/metric DB writes

Pure structural move: every symbol relocates verbatim, function-local
imports stay lazy, and __init__.py re-exports health_command and
_render_refactoring_targets so the public import path is unchanged.
